refactor: rename tasks-board provider to context
Replaced `provider` with `context` in tasks-board components for consistency with naming conventions. Updated all imports, exports, and references to reflect this change.
