+++ apps/nextjs/src/components/features/tasks/tasks-board/tasks-board-list-item/provider/provider.tsx
@@ -1,6 +1,6 @@
 import { memo, type PropsWithChildren } from 'react';
 import { Context as InputContext } from './input-provider';
-import { Context as ListItemContext } from './list-item-provider';
+import { Context as ListItemContext } from './list-item-context';
 
 type Props = PropsWithChildren<{
   taskId: string;
